[STORY-P00-001] fix(p00): reconcile engineering baseline and continuity
Task-ID: TASK-P00-001
Session-ID: SES-20260716T232809Z-B4A980AF
Checkpoint-ID: CP-SES-20260716T232809Z-B4A980AF-0007
Tests: PASS:alert-sink,android-build,backend-local-verify,backend-postgresql-17.10,continuity-doctor,continuity-self-test,p00-matrix-unit,project-doctor
CR: CR-0003,CR-0005,CR-0006
Story-ID: STORY-P00-001

+++ services/backend/platform/src/main/java/cc/orbexa/hhy/platform/release/AppVersionPolicyView.java
@@ -2,5 +2,13 @@ package cc.orbexa.hhy.platform.release;
 
 import java.time.Instant;
 
-public record AppVersionPolicyView(String platform, int latestVersionCode, String latestVersionName, String updateType,
-                                   String downloadUrl, String sha256, String releaseNotes, int minSupportedVersionCode, Instant serverTime) { }
+public record AppVersionPolicyView(
+        String platform,
+        long latestVersionCode,
+        String latestVersionName,
+        String updateType,
+        String downloadUrl,
+        String sha256,
+        String releaseNotes,
+        long minSupportedVersionCode,
+        Instant serverTime) { }
